rename hookspec module
