+++ napari/plugins/hook_specifications.py
